backend/requirements.txt
alembic==1.6.2
bcrypt==3.2.0
cffi==1.15.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Migrate==2.7.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonify==0.5
make-response==1
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
mysqlclient==2.0.3
protobuf==3.16.0
pycparser==2.21
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.14
Werkzeug==1.0.1
